refactor: move cake day feature to its own directory
