+++ application/src/main/java/org/togetherjava/tjbot/features/Features.java
@@ -6,7 +6,6 @@ import org.togetherjava.tjbot.config.Config;
 import org.togetherjava.tjbot.config.FeatureBlacklist;
 import org.togetherjava.tjbot.config.FeatureBlacklistConfig;
 import org.togetherjava.tjbot.db.Database;
-import org.togetherjava.tjbot.features.basic.CakeDayRoutine;
 import org.togetherjava.tjbot.features.basic.MemberCountDisplayRoutine;
 import org.togetherjava.tjbot.features.basic.PingCommand;
 import org.togetherjava.tjbot.features.basic.RoleSelectCommand;
@@ -16,6 +15,7 @@ import org.togetherjava.tjbot.features.bookmarks.BookmarksCommand;
 import org.togetherjava.tjbot.features.bookmarks.BookmarksSystem;
 import org.togetherjava.tjbot.features.bookmarks.LeftoverBookmarksCleanupRoutine;
 import org.togetherjava.tjbot.features.bookmarks.LeftoverBookmarksListener;
+import org.togetherjava.tjbot.features.cakeday.CakeDayRoutine;
 import org.togetherjava.tjbot.features.chatgpt.ChatGptCommand;
 import org.togetherjava.tjbot.features.chatgpt.ChatGptService;
 import org.togetherjava.tjbot.features.code.CodeMessageAutoDetection;
